# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/masjids/{id}": {
      "get": {
        "operationId": "AdminMasjidsController_getMasjid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Requires permission: platform.masjids.manage"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get masjid details by id",
        "tags": [
          "Admin Masjids"
        ]
      }
    },
    "/admin/masjids/{id}/status": {
      "patch": {
        "description": "REJECTED and SUSPENDED require a reason (400 VALIDATION_ERROR). Unknown id: 404 MASJID_NOT_FOUND.",
        "operationId": "AdminMasjidsController_updateMasjidStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMasjidStatusDto"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Requires permission: platform.masjids.manage"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update masjid status",
        "tags": [
          "Admin Masjids"
        ]
      }
    },
...
    "/masjid-requests": {
      "post": {
        "description": "Public endpoint. No bearer token is required.",
        "operationId": "MasjidRequestsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMasjidRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Masjid request created successfully"
          }
        },
        "summary": "Submit a public masjid registration request",
        "tags": [
          "Masjid Requests"
        ]
      },
      "get": {
        "operationId": "MasjidRequestsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "jama",
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "example": "4e0798d2-3fd3-4caa-9966-9f85c96f8b2f",
              "type": "string"
            }
          },
          {
            "name": "masjidName",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Jama Masjid",
              "type": "string"
            }
          },
          {
            "name": "locality",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Nighasan",
              "type": "string"
            }
          },
          {
            "name": "district",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Bhopal",
              "type": "string"
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "schema": {
              "example": "India",
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Madhya Pradesh",
              "type": "string"
            }
          },
          {
            "name": "requesterPhone",
            "required": false,
            "in": "query",
            "schema": {
              "example": "9876543210",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Masjid requests retrieved successfully"
          },
          "403": {
            "description": "Requires permission: platform.masjid_requests.manage",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "Forbidden resource",
                    "errorCode": "FORBIDDEN"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all masjid requests with filters",
        "tags": [
          "Masjid Requests"
        ]
      }
    },
    "/masjid-requests/track": {
      "post": {
        "description": "Public endpoint. No bearer token or tracking token is required. Returns { items } with the 20 newest applications for the phone; this list is not paged (no meta).",
        "operationId": "MasjidRequestsController_track",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackMasjidRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Applications fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Applications fetched successfully",
                    "data": {
                      "items": [
                        {
                          "masjidName": "Jama Masjid",
                          "status": "PENDING",
                          "imamName": "Abdul Rahman",
                          "requestedAt": "2026-06-19T00:00:00.000Z",
                          "reviewedAt": null
                        }
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Track public masjid registration requests by requester phone",
        "tags": [
          "Masjid Requests"
        ]
      }
    },
    "/masjid-requests/{id}": {
      "get": {
        "description": "Same shape as an item of GET /masjid-requests.",
        "operationId": "MasjidRequestsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "4e0798d2-3fd3-4caa-9966-9f85c96f8b2f",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Masjid request retrieved successfully"
          },
          "403": {
            "description": "Requires permission: platform.masjid_requests.manage",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "Forbidden resource",
                    "errorCode": "FORBIDDEN"
                  }
                }
              }
            }
          },
          "404": {
            "description": "MASJID_REQUEST_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "Forbidden resource",
                    "errorCode": "FORBIDDEN"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one masjid request",
        "tags": [
          "Masjid Requests"
        ]
      }
    },
    "/masjid-requests/{id}/status": {
      "patch": {
        "description": "Only a PENDING request can change. Already decided: 409 MASJID_REQUEST_ALREADY_APPROVED / MASJID_REQUEST_ALREADY_REJECTED (also for the loser of two simultaneous decisions). Imam or committee member in another masjid: 409 USER_IN_ANOTHER_MASJID. New imam whose email belongs to another account: 409 EMAIL_ALREADY_EXISTS.",
        "operationId": "MasjidRequestsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "4e0798d2-3fd3-4caa-9966-9f85c96f8b2f",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMasjidRequestStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Masjid request status updated successfully"
          },
          "403": {
            "description": "Requires permission: platform.masjid_requests.manage",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "Forbidden resource",
                    "errorCode": "FORBIDDEN"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "Forbidden resource",
                    "errorCode": "FORBIDDEN"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve or reject a masjid request",
        "tags": [
          "Masjid Requests"
        ]
      }
    },
...
  },
  "info": {
    "title": "Masjid Core API",
    "description": "Backend for the masjid management system. Every masjid route needs a permission; see src/access/permissions.ts.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token from /auth/login/verify-otp"
      }
    },
    "schemas": {
...
      "UpdateMasjidStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "SUSPENDED"
            ],
            "example": "APPROVED"
          },
          "reason": {
            "type": "string",
            "example": "Duplicate registration"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "ImamDetailsDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Imam Name",
            "maxLength": 150
          },
          "email": {
            "type": "string",
            "example": "imam@example.com",
            "maxLength": 255
          },
          "phone": {
            "type": "string",
            "example": "9876543210",
            "maxLength": 20
          },
          "address": {
            "type": "string",
            "example": "Single line address",
            "maxLength": 500
          }
        },
        "required": [
          "name",
          "phone"
        ]
      },
      "CommitteeMemberDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Committee Member Name",
            "maxLength": 150
          },
          "phone": {
            "type": "string",
            "example": "9876543211",
            "maxLength": 20
          },
          "fatherName": {
            "type": "string",
            "example": "Abdul Kareem",
            "maxLength": 150
          },
          "age": {
            "type": "number",
            "example": 42,
            "minimum": 1,
            "maximum": 120
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ],
            "example": "MALE"
          }
        },
        "required": [
          "name",
          "phone",
          "fatherName",
          "age",
          "gender"
        ]
      },
      "CreateMasjidRequestDto": {
        "type": "object",
        "properties": {
          "requesterName": {
            "type": "string",
            "example": "Person Name",
            "minLength": 2,
            "maxLength": 150
          },
          "requesterPhone": {
            "type": "string",
            "example": "9876543210",
            "maxLength": 20
          },
          "requesterEmail": {
            "type": "string",
            "example": "person@example.com",
            "maxLength": 255
          },
          "masjidName": {
            "type": "string",
            "example": "Jama Masjid",
            "minLength": 2,
            "maxLength": 150
          },
          "country": {
            "type": "string",
            "example": "India",
            "maxLength": 100
          },
          "district": {
            "type": "string",
            "example": "District",
            "maxLength": 100,
            "description": "Required when country is India or IN"
          },
          "locality": {
            "type": "string",
            "example": "Nighasan",
            "maxLength": 100,
            "description": "City / Village / Town"
          },
          "state": {
            "type": "string",
            "example": "State",
            "maxLength": 100
          },
          "address": {
            "type": "string",
            "example": "Address",
            "maxLength": 500
          },
          "contactNo": {
            "type": "string",
            "example": "9876543210",
            "maxLength": 20
          },
          "description": {
            "type": "string",
            "example": "Short description",
            "maxLength": 1000
          },
          "welcomeMsg": {
            "type": "string",
            "example": "Assalamu Alaikum",
            "maxLength": 500
          },
          "imamName": {
            "type": "string",
            "example": "Imam Name",
            "maxLength": 150
          },
          "imamPhone": {
            "type": "string",
            "example": "9876543210",
            "maxLength": 20
          },
          "imamEmail": {
            "type": "string",
            "example": "imam@example.com",
            "maxLength": 255
          },
          "imamAddress": {
            "type": "string",
            "example": "Single line imam address",
            "maxLength": 500
          },
          "imamFatherName": {
            "type": "string",
            "example": "Abdul Kareem",
            "maxLength": 150
          },
          "imamAge": {
            "type": "number",
            "example": 50,
            "minimum": 1,
            "maximum": 120
          },
          "imamGender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ],
            "example": "MALE"
          },
          "imam": {
            "deprecated": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ImamDetailsDto"
              }
            ]
          },
          "committeeMembers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommitteeMemberDto"
            }
          }
        },
        "required": [
          "requesterName",
          "requesterPhone",
          "masjidName",
          "country",
          "locality",
          "state",
          "address",
          "imamName",
          "imamPhone",
          "imamAddress",
          "imamFatherName",
          "imamAge",
          "imamGender",
          "committeeMembers"
        ]
      },
      "TrackMasjidRequestDto": {
        "type": "object",
        "properties": {
          "requesterPhone": {
            "type": "string",
            "example": "+919876543210"
          }
        },
        "required": [
          "requesterPhone"
        ]
      },
      "UpdateMasjidRequestStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "REJECTED"
            ],
            "example": "APPROVED"
          },
          "reason": {
            "type": "string",
            "example": "Verified successfully",
            "maxLength": 500
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}